Django-Middleware-0x03/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
drf-nested-routers==0.94.2
idna==3.10
parameterized==0.9.0
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
